Phucnt - FixDAO, import Template to PRJ (#21)
* delete AccountFactory

* Add Login SQL script and checkLogin in AccountDAO + change the visual of register and login page

* implements CRUD method and access to DB in AccountDAO and change param in DAO

* add List discount id to User

* aadd Login Controller to prj

* fix DAO

* import shop template and css to prj

+++ src/java/com/model/genres/GenreDAO.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 public class GenreDAO implements DAO<Genre> {
 
     @Override
-    public Optional<Genre> get(long id) {
+    public Optional<Genre> get(int id) {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
